Squashed commit of the following:

commit d10d2ee40c49afc0e3e4b79f596806c4b4012fa3
Author: Konrad Kopciuch <konrad.kopciuch@relativity.com>
Date:   Thu Jun 25 12:49:26 2020 +0200

    REL-435278 code review feedback

    - updated readme.md with description of tests variables
    - added new tests variables to TestParameters-Template.reg
    - use underscore for private fields

commit 803a2729803f4cf2c0341367974d2842153743cd
Author: Konrad Kopciuch <konrad.kopciuch@relativity.com>
Date:   Thu Jun 25 11:24:24 2020 +0200

    REL-435278 fixed issue in SQLDataComparer

    Removed WHERE clause which was causing issues and is not longer necessary.

commit 3cc8d7e0fe37e200a929c7169bd64e8337106125
Merge: c21808e1a f6e4a1872
Author: Konrad Kopciuch <konrad.kopciuch@relativity.com>
Date:   Thu Jun 25 11:18:27 2020 +0200

    Merge remote-tracking branch 'origin/develop' into REL-435278-integrate-tests-with-SQLDataComparer

commit c21808e1a78e86405ecfc02778d19c287fc499a9
Author: Konrad Kopciuch <konrad.kopciuch@relativity.com>
Date:   Wed Jun 24 15:44:06 2020 +0200

    REL-435278 tests can generate input for SQLDataComparer tool

    - extracted SQLDataComparer configuration to xml file from app.config
    - passing SqlServer instance in credentials as parameters to SQLDataComparer
    - tests decorated with [UseSqlComparer] generate input for SQLDataComparer
    - decorated two load tests with that attributed and prepared comparer configuration for those tests
    - I replaced app.config in integration tests projects by a single app.config in test framework. Added that app.config as a link

+++ Source/Relativity.DataExchange.TestFramework/IntegrationTestHelper.cs
@@ -147,42 +147,21 @@ namespace Relativity.DataExchange.TestFramework
 				return;
 			}
 
-			WorkspaceHelper.DeleteTestWorkspace(parameters, Logger);
-			string database = $"EDDS{parameters.WorkspaceId}";
-			if (parameters.SqlDropWorkspaceDatabase && parameters.WorkspaceId > 0)
+			if (parameters.DeleteWorkspaceAfterTest)
 			{
-				try
+				WorkspaceHelper.DeleteTestWorkspace(parameters, Logger);
+				if (parameters.SqlDropWorkspaceDatabase && parameters.WorkspaceId > 0)
 				{
-					SqlConnectionStringBuilder builder = GetSqlConnectionStringBuilder(parameters);
-
-					SqlConnection.ClearAllPools();
-					using (SqlConnection connection = new SqlConnection(builder.ConnectionString))
-					{
-						connection.Open();
-						using (SqlCommand command = connection.CreateCommand())
-						{
-							command.CommandText = $@"
-IF EXISTS(SELECT name FROM sys.databases WHERE name = '{database}')
-BEGIN
-	ALTER DATABASE [{database}] SET SINGLE_USER WITH ROLLBACK IMMEDIATE;
-	DROP DATABASE [{database}]
-END";
-							command.CommandType = CommandType.Text;
-							command.ExecuteNonQuery();
-							Logger.LogInformation("Successfully dropped the {DatabaseName} SQL workspace database.", database);
-							Console.WriteLine($"Successfully dropped the {database} SQL workspace database.");
-						}
-					}
+					DropWorkspaceDatabase(parameters);
 				}
-				catch (Exception e)
+				else
 				{
-					Logger.LogError(e, "Failed to drop the {DatabaseName} SQL workspace database.", database);
-					Console.WriteLine($"Failed to drop the {database} SQL workspace database. Exception: " + e);
+					Logger.LogInformation("Skipped dropping the SQL workspace {workspaceId} database", parameters.WorkspaceId);
 				}
 			}
 			else
 			{
-				Logger.LogInformation("Skipped dropping the {DatabaseName} SQL workspace database.", database);
+				Logger.LogInformation("Skipped deleting the workspace {workspaceId}.", parameters.WorkspaceId);
 			}
 		}
 
@@ -366,6 +345,40 @@ END";
 			{
 				string currentPath = new FileInfo(Assembly.GetExecutingAssembly().Location).DirectoryName;
 				parameters.SqlProfilingReportsOutputPath = Path.GetFullPath(Path.Combine(currentPath, sqlProfilingReportsOutputPathValue));
+			}
+		}
+
+		[System.Diagnostics.CodeAnalysis.SuppressMessage("Microsoft.Design", "CA1031:DoNotCatchGeneralExceptionTypes", Justification = "We can swallow exception in that case.")]
+		private static void DropWorkspaceDatabase(IntegrationTestParameters parameters)
+		{
+			string database = $"EDDS{parameters.WorkspaceId}";
+			try
+			{
+				SqlConnectionStringBuilder builder = GetSqlConnectionStringBuilder(parameters);
+
+				SqlConnection.ClearAllPools();
+				using (SqlConnection connection = new SqlConnection(builder.ConnectionString))
+				{
+					connection.Open();
+					using (SqlCommand command = connection.CreateCommand())
+					{
+						command.CommandText = $@"
+IF EXISTS(SELECT name FROM sys.databases WHERE name = '{database}')
+BEGIN
+	ALTER DATABASE [{database}] SET SINGLE_USER WITH ROLLBACK IMMEDIATE;
+	DROP DATABASE [{database}]
+END";
+						command.CommandType = CommandType.Text;
+						command.ExecuteNonQuery();
+						Logger.LogInformation("Successfully dropped the {DatabaseName} SQL workspace database.", database);
+						Console.WriteLine($"Successfully dropped the {database} SQL workspace database.");
+					}
+				}
+			}
+			catch (Exception e)
+			{
+				Logger.LogError(e, "Failed to drop the {DatabaseName} SQL workspace database.", database);
+				Console.WriteLine($"Failed to drop the {database} SQL workspace database. Exception: " + e);
 			}
 		}
 	}
